Updated openstack/openstack Project: openstack/keystone  68e94ff979040370ec9e9310e8cc1de3382037a1
Add entrypoint for mapped auth method

The keystone federation documentation says to use the 'mapped' method
instead of the 'saml2' method[1]. However, that entrypoint was never
added, so trying to use it resulted in keystone not being able to load
its wsgi scripts and failing. This patch adds the entrypoint so that
keystone won't explode.

[1] http://docs.openstack.org/developer/keystone/federation/federated_identity.html#configuring-federation-in-keystone

Change-Id: I6dd8c219765728ecf01b50f35c3c0da3943f13b0