Updated openstack/openstack Project: openstack/python-openstackclient  351bfb79154777dca1bc117e9383606f16b02353
Fix post and periodic jobs

The usage of zuul-cloner in tox_install breaks post and periodic jobs.
Update the script, it does not need to handle unconstrainted installs
anymore (see tox.ini). There's no need to set ZUUL_BRANCH explicitely
- it's set via the environment and branch is passed in as well.

Note that this script is needed for the edit-constraints call, add a
comment.

Change-Id: I0077c986a17d6bb92791474e03d1e77776e9382f
Closes-Bug: #1615430