Updated openstack/openstack Project: openstack-infra/project-config  062d2bd10b92ebd25a455fb02a2da1f3df9e2227
Add os_watcher to OpenStack-Ansible

This patch adds a new pre-seeded git repository to OpenStack-Ansible
for deploying Watcher.

The os_watcher repository has its own core team with the general
openstack-ansible core team included with the appropriate rights to
ensure that they can act as a back-stop in case of any issues.

Needed-By: I0fdf31d5234bd7fea4114ba02c8d5ac9f164b81c
Change-Id: I54eee14c4030dc149585b7113483510fdbcaf55d