Updated openstack/openstack Project: openstack-infra/tripleo-ci  2df9fdc9acd4d8b7a91c6b9f5828296ee17ee4dd
Use separated SSL endpoint environment file

We shouldn't be maintaining the EndpointMap in two different repos.
This CR uses the standalone EndpointMap environment file that is now
present in t-h-t. We then add a separate endpoint list environment
file to the test-environments directory since we need it for liberty
and mitaka.

Change-Id: I967b9ee9fd5145b447fce8255ba27def458b0e91
Depends-On: I38eabb903b8382e6577ccc97e21fbb9d09c382b3