Updated openstack/openstack Project: openstack/keystone  a4451658153f3abccfac81ac1f052eef5e46c70b
Shadowing a nonlocal_user incorrectly creates a local_user

This patch fixes a bug where when shadowing a nonlocal_user (LDAP,
custom driver) it also incorrectly creates a local_user. The error is
related to hybrid properties and calling the class from_dict method,
which set the local_user attributes.

Change-Id: I6e69cce5f337a330f2531ff71db3e931b785271c
Closes-Bug: #1615000