Updated openstack/openstack Project: openstack/trove-integration  c637e5ffc2a29dabbf17decb70f6b3a05ab4b8ae
Turn off Neutron until supporting changes land

There are a number of changesets that need to be merged
before Trove can use Neutron successfully.  Since two
of these are in devstack we'll need to wait to see if they
are willing to accept the changes before we can start using
it by default.

See: https://review.openstack.org/#/c/356026
     https://review.openstack.org/#/c/356763
     https://review.openstack.org/#/c/356701

Note that the 2nd changeset (devstack) could be rolled into the
Trove plugin (3rd changeset) by copying the relevant functions
(not ideal, but doable) however the 1st changeset (devstack) is
fairly critical and would be difficult to implement outside
the devstack Neutron l3 plugin.

Also changed the logic for the Neutron client (and others)
such that the code is pulled from git if the project is enabled,
and made sure *_BRANCH and PROJECT_BRANCH were honoured on the plugins.

Change-Id: I764f15514a40926173b6c9194ddbf24c6c1f999a