Updated openstack/openstack Project: openstack-infra/release-tools  55c0dcb126ae5c870e506b4370402a8a27b21452
fix announce.sh for projects with setup_requires

The XStatic projects can't use pbr because they need four digit
versions. They use setuptools-scm to collect their file list instead,
but that is not pre-installed on the nodes where the announce script
runs. Rather than deal with that explicitly, run a setuptools command to
cause all of the needed things to be installed in a way that will log
the output but not include it in the outgoing email body.

Change-Id: I8ce14b19dc13bad6260a3e32f14e06f5f5fc5df4
Signed-off-by: Doug Hellmann <doug@doughellmann.com>