Updated openstack/openstack Project: openstack/trove  afbd993b891c841e592bf4ee1ec9d55264925ea5
Extend guest agent call timeouts

The gate is horribly overloaded and as a result our calls to the
guest are timing out during scenario tests.  I've been told that
the community won't support fixing the tests, so the only option
is to break production code.  This change makes the high timeout
10 minutes, which is now the amount of time the user will need
to wait to learn of an error in an asyncronous call to the guest.

Change-Id: I66d97dc4a60a1d2afe12d27d167ca426b51074ca