Updated openstack/openstack Project: openstack/nova  7a3d1b042c878a64d9d57365f42f5974bebf06f5
Hyper-V: update live migrate data object

The Hyper-V driver does not properly handle shared storage at the moment.

In order to address this, and ensure that proper cleanup is performed
in case of failed live migrations, we'll need to update the
HyperVLiveMigrateData object and make sure that the compute manager
properly uses it.

This change is part of a bigger series, attempting to fix HyperV shared
storage related issues.

Change-Id: I1beb56530134c83b8359ac137cdffc37a86ec8cc
Partial-Bug: #1565895