Updated openstack/openstack Project: openstack/heat  d850605da36deab2621b3c766a25efdd47a9d05a
Load resources from DB for resource list

Currently, when a user makes an API call to list resources, they were
retrieved from the currently active template associated with the
stack. This worked in the legacy engine where no concurrent updates
where possible. With convergence and thus concurrent updates, a stack
is allowed to have resources of previous traversals and still continue
creating new resources. Therefore, relying on the template to list
resources won't exactly match the state in the database.

For example, on deletes where we update with an empty template,
currently, the stack parses the empty template searching for
resources. Doing a `resource-list` when the stack is in a state of
DELETE_IN_PROGRESS shows an empty list of resources and therefore not
matching the state in the database.

This change makes iter_resources always call _find_filtered_resources
which builds all its resources from a database query.

Change-Id: Ibe87a773c38efb6d4865fd3a1dbd079972dd8be4
Closes-Bug: #1523748
Closes-Bug: #1301320