Updated openstack/openstack Project: openstack-infra/puppet-gerrit  5c44d09dc717068f7ea5ee0ff0d2319863991817
Create index on fresh install

A fresh install of gerrit was refusing to start with

  com.google.inject.ProvisionException: Unable to provision, see the following errors:

  1) No index versions ready; run Reindex

Because we pre-populate some of the directories in review_site, gerrit
init thinks it is in an upgrade situation and does not create or touch
the index.  Thus manually create the index when we are initalising a
new site.

Also, the gerrit-reindex doesn't have to subscribe to the
gerrit-initial-init; it should only be triggered on updates if
offline-indexing is wanted/required.

Change-Id: I98ce9ff87570a86e9b58b38d0605313ee5ee2758