Updated openstack/openstack Project: openstack/keystone  3de1713f2de84c1549f00f7e6df6fd47707c4a3d
Fix credential update to ec2 type

It was possible to create a credential without providing a project_id
and later updating it to the ec2 type.

This patch fixes the issue by adding a manual checking in the
manager layer since it needs to check the old credential contents
prior failing the request.

Change-Id: I1eb28a46c89e17d9c990cc798867d1a59714fe5f
Closes-Bug: #1613466