Updated openstack/openstack Project: openstack/cinder  afd5fc5f85c407bd33bbb8d6ef82b0a93e4990b1
Change NFS driver to not throw exception without nfs shares file

Currently, the NFS driver throws an exception if there isn't
at least an empty file in existence for the default
/etc/cinder/nfs_shares file.  This, however, isn't
consistent with the documentation or the way that
the driver actually works.

If the nas_host and nas_share_path are configured the
nfs_shares file is not used, so throwing an exception is
not necessary.  All the information for setting up the
NFS shares is available in the config file.

This change updates the NFS driver to only throw an
exception if the configured nfs_shares file doesn't
exist and nas_host and/or nas_share_path are missing.

Unit tests are included with the change.

Closes-bug: #1602459

Change-Id: Ic2697a25e5ff20c84d946515bde32a4697d326c6