Updated openstack/openstack Project: openstack/nova  8dbf0529cbd001761f53a72a6640a5c90874e9e0
Add a Usage and UsageList object

These encapsulate the Usage information associated with the
inventories of a resource provider. A single UsageList holds
all the Usage for all resource classes provided by that provider.
Just resource class and usage.

Putting this in place to make further headway on the placement
API, which needs to represent Usage info in this form.

This change intentionally tries to keep the new objects
relatively lightweight.

Partially-Implements: blueprint generic-resource-pools
Change-Id: Ie5eddae4d0cb51bccc815a932a3eefcd5d9ea687