Updated openstack/openstack Project: openstack-infra/tripleo-ci  52ada87af9e53895138312e3a2390768146fb9f8
Deploy minimal services in multinode job

Deploy less services in the multinode job, only the required ones. Since
this job is in the gate we should strive to make it as efficient as
possible and focus on testing the deployment framework itself, as least
initially. We could always add back or add additional serivces later.

Change-Id: I1bb20c279e50b5556f3c67ab6d2be72a8133a2a6