Updated openstack/openstack Project: openstack-infra/project-config  5088ba97a3656b262c3b34e312da9ecaae3d6d14
TripleO scenario001 experimental job

TripleO is currently investigating how to scale the number of services
without overload the CI jobs.

Puppet CI had the same challenge a few months ago and solved it by
running multiple scenarios covering a different set of services in CI.

TripleO is looking for something similar.

This patch is a start of something that:
- create experimental job called scenario001-multinode
- run the job only if sahara puppet or hiera files are modified.

Note: it's called scenario001 so we can document what scenario001 is
doing. Putting the services in the scenario name won't help us, because
we apply complex regex on job names, etc. Let's use scenarioXXX for now
and see how it goes. It's also consistent with Puppet CI jobs.

Change-Id: I2b626da3b5cfd044b76bcb3741cd8d7426b0293a