Updated openstack/openstack Project: openstack/ironic-python-agent  95f31be108f6f1a259d64535967d796a25e4d7cd
Re-use API client for Heartbeat operations

Re-initialization of the API client causes the identification
of an older ironic API to be forgotten.

As such, we need to re-use the API client from the initial lookup
for the heartbeat.

Tested with current master branch and stable/mitaka ipa
in noauth mode, and was able to deploy without any issues.

Closes-Bug: #1616461
Change-Id: I3ae1e2ce83f614b5b936a3cbb1b9587c6c55a89a