Updated openstack/openstack Project: openstack/tripleo-heat-templates  d7035cf80c1d79fad555e125799805c964c154a9
Fix check for MariaDB upgrade manual switch off

The MySqlMajorUpgrade parameter has validation on it allowing only
values yes/no/auto, however in the script we checked for '0' instead of
'no', which means the only effective values were yes/auto. This is now
fixed to allow switching the migration off.

Change-Id: I5d64734894c6bfd9003ad643f3747e34e62465cc
Closes-Bug: #1616429