Updated openstack/openstack Project: openstack/keystone  f46185feb70ede6df33155cb8c13df5db4b8dbe7
Update `href` for keystone extensions

Since the API docs have been moved to api-ref, and the docs maintained
in keystone-specs repo have been moved to attic. The href attribute
that points to old `specs.openstack.org` will not be a valid link, this
patch update them to point to the new site.

Change-Id: Idc4995509d262c74ce6d49c6c18b6452fae186d6