Updated openstack/openstack Project: openstack/neutron  b5e2902b1739dc2e5dd54210ba8df3fa2767910d
Fix bug in L3 agent extension manager

In review [1] it was noticed by a reviewer that the wrong object was
being consulted in the L2 agent extension manager when checking to see
if the handle_port and delete_port methods were implemented.  Since the
L3 agent extension manager has the same code - which had already
merged - it will require the same fix.

[1] https://review.openstack.org/#/c/355220/3..4/neutron/agent/l2/l2_agent_extensions_manager.py

Change-Id: I2793474abd46dc2553b70bccad3ee25258a3707f