Updated openstack/openstack Project: openstack/tripleo-heat-templates  6d12839f3655f8774ad227bfa4f3cf4d7b45acc5
Allow a fixed IP to be allocated to RedisVirtualIP

The overcloud stack now contains a parameter RedisVirtualFixedIPs which will
allow people to specify the IP address they would like to be assigned as a vip
used by Redis. This will allow people full control over what IPs are used as
VIPs during a deployment.

Change-Id: Ib24e62847c4ac43fcd2a09bc84b4a76e992add4b