Updated openstack/openstack Project: openstack/nova  1ce9613f874098795db34aa72979950911413787
Remove TODO for service version caching

A TODO was added when enabling service version caching for the API but
it was based on an assumption that the caching was only used for the
"Get me a network" feature. Now that a second feature "Cellsv2" is going
to benefit from that caching let's remove the note about removing the
caching.

Change-Id: Ibd1676ae617de7b260e8d919e1c57e2eec4f5373