Updated openstack/openstack Project: openstack/heat  815fe6452e434772d683ce4545ec6b2a0907e157
Add a Macro class to heat.engine.function

A Macro differs from a function in that it modifies the syntax tree
directly, so that the result of inserting a macro in the tree might not be
just different output, but actually be based on resolving different
functions. A macro can remove syntax from the tree, modify it or fabricate
it from whole cloth prior to parsing.

The proposed If 'function' for implementing conditionals is a macro.
Arguably the Repeat function should be implemented as a macro also.

Change-Id: I2c5afe31d509ffedc9e0b2ef237f52d04687e0be
Partially-Implements: blueprint support-conditions-function