Updated openstack/openstack Project: openstack/tripleo-heat-templates  2903a9e4cdb55628ecdd76557e9c7814827d04ab
Move resource registry override to enable-tls.yaml

It makes more sense for the enable-tls.yaml file to contain the
resource registry override, since it contains parameters that are
actually used there. Also, this allows us to reuse the
tls-endpoints-public-* files for other methods of enabling TLS (such
as with certmonger).

Change-Id: I98c63d0007e61968c0490a474eddb42548891fa6