Updated openstack/openstack Project: openstack/neutron  b70803a14435160d05f3a74cddf584f359c6c4cc
Make OVS and LinuxBridge trunk drivers' is_loaded() property more robust

If ML2 is not in use, CONF.ml2.mechanism_drivers is not a registered
option and thus loading these drivers leads to a NoSuchOptError error.
This patch ensures that the exception is handled and the loading state
of the OVS and Linuxbridge driver is set to False.

Partial-implements: blueprint vlan-aware-vms

Change-Id: I0a6994b9248e878e314661f235ceadb7ce9bbbff