Updated openstack/openstack Project: openstack/cinder  f3d62bbb18a48675d329755237a170e10333495c
Retype issue in Kaminario Cinder drivers

The value of volume.replication_status for a replicated volume is
'disabled' before 'failover-host' and it is changed to "failed-over"
after 'failover-host' command if active_backend_id is not equal to
secondary array IP. It is changed to 'enabled' after another
'failover-host' command if active_backend_id is equal to secondary
array IP(which is failback scenario).

After failback, if replication is removed from replicated volume,
then retype 'dedup without replication'<->'nodedup without replication'
with '--migration-policy on-demand' is not working since
value of volume.replication_status is enabled.

Change-Id: I193572532a36eaea5d826cff76ca9f98df86dc50
Closes-Bug: #1616329
Co-Authored-By: VenkataKrishna Reddy<Venkata.Krishna.ctr@kaminario.com>
Co-Authored-By: Lakshman<Lakshmi.Narayana.ctr@kaminario.com>