Updated openstack/openstack Project: openstack/tripleo-specs  1b476d3edbf7158b1026f592e795ceaf0705acfa
tripleo-sriov

This spec file covers the specifications to add SR-IOV functionality
in tripleo.

SR-IOV extends the PCI Express specification and allows a PCIe device
to appear to be multiple separate physical PCIe devices. SR-IOV
provides one or more Virtual Functions (VFs) and a Physical
Function(PF).  These VF’s could be attached to VMs like a dedicated
PCIe device, which boosts the networking performance considerably.

This feature is supported in OpenStack, but the setup and
configuration part is manual and it is not supported in TripleO.

Co-Authored-By: Karthik Sundaravel <ksundara@redhat.com>

Change-Id: I9f336a0c80575558575c85b6628954d3c8422610