Updated openstack/openstack Project: openstack/ossa  5517e5b17a6d745369b4e16dff1634bc85493aeb
Add "Invalid" status to OSSA Task status

"Invalid" should be used when the reported issue is not a bug or
is marked as invalid by the project as it is not a bug/vulnerability.
This is distinct from "Won't Fix" as it indicates there is something
to the report, but nothing for the VMT to do.

Change-Id: I72598c0954a17fd26a4f12513ff28fc2625c374e