Updated openstack/openstack Project: openstack/python-swiftclient  9cdeedab7dea2d0b224c7347d9fe9a2268ae167a
Use mock patch to handle get_auth_keystone

Setting c.get_auth_keystone = fake_get_auth_keystone is setting a new
method on the module. This information is not reset between test runs
and therefore has the potential to corrupt other tests.

Use mock patch so that the mock is reset after the test is complete.

Change-Id: Ifb9ba72634cf477c72dda080b5aed8f8e3a7ac89