Updated openstack/openstack Project: openstack/nova  7500bef94f526a82392400415f07f744700324a9
Revert "Optional separate database for placement API"

This reverts commit 1b5f9f8203c90fe447d33c89f238104026052d1e.

On IRC I we decided that we agreed no migrations should be placement
specific, we should just use the API table migrations to generate the
schema for both DBs.

There is also a separate debate around the alias for the aggregates
table, but that is not really a reason to revert, its just things in
here that will need rework.

Change-Id: I275945aee9d9be8e35d6ddc05515df39d559457a