Updated openstack/openstack Project: openstack/oslo.vmware  6bf58463e0f11cd0a3a868d74ef7db3973da2912
Pass connection timeout so that invoke_api will not wait forever

invoke_api does not throw any exception when interface goes down and
application loses connectivity with vCenter because default timeout
is set to None which is equivalent to infinity.

This patch set allows any application to pass connection timeout so that
it will wait for the given timeout and then throw an exception.

Closes-Bug: #1564839

Change-Id: I69b099b7caaa4fb0e0067cb83b9468d4aa901cc7
Co-Authored-By: Raghuveer Shenoy <rshenoy@hp.com>