Updated openstack/openstack Project: openstack-infra/project-config  d788034b1e4e65d233b300e5d4595574e5773d12
Add os_trove to OpenStack-Ansible

This patch adds a new pre-seeded git repository to OpenStack-Ansible
for deploying Trove.

The os_trove repository has its own core team with the general
openstack-ansible core team included with the appropriate rights to
ensure that they can act as a back-stop in case of any issues.

Needed-By: I0fdf31d5234bd7fea4114ba02c8d5ac9f164b81c
Change-Id: I5146f5a4c1bd039d5a12011fee4f3bb3a19145c4