Updated openstack/openstack Project: openstack-infra/project-config  a7b233d2de2ab54ea15f4e81ad96e3f59c2e9655
Drop postgresql jobs in gate

postgresql is used by < 1% of our community, and the tempest dsvm
tests are actually an extremely poor place to validate pg, because the
differences are mostly not addressable at this level. There are better
uses of our CI resources.

Depends-On: I141bb7a44b5c34620479dd56496b21e0d95d9937
(which ensures metadata server is tested)

Change-Id: I5a4163e095ce132c3f9615328fe4b4e96d7936ad