Updated openstack/openstack Project: openstack/tripleo-heat-templates  3f5a6b1a820f9cbb1943c628348f93b9f8a6b22b
Enable usage of FQDNs for the endpoints

The endpoint map has the capability of using the cloud's name for the
endpoint. This is broken, however, since this has the problem that we
only take into account the overcloud's external endpoint name, which
we then cannot use if we have network-isolation enabled, which is
the most common use-case for real deployments.

So this change proposes the following:
* The external endpoint is still CloudName.
* We can now set different (or the same if we want) names for the
  different VIPs of the network.
* Using CLOUDNAME for the endpoint map will get a name for the
  appropriate network.

bp tls-via-certmonger

Change-Id: I3e7144653f0a1d783d87e6f638304b297f718929