Updated openstack/openstack Project: openstack-dev/pbr  8d64858523c68b4111e98b14b01f615d79ee4048
Add more words to a confusing error message

When setup.cfg has a different value than pbr.version.VersionInfo(),
then the version call is unable to look up the right value in the
pkg_resources metadata. However, nothing in the current error message
communicates this to the user, which leads to them not being able to
look in all the right places for the possible error.

Change-Id: I31c4e0e2ed15986ba8274a0c5270a9d214ac4b48