Updated openstack/openstack Project: openstack/cinder  852176997ae2b8c4f4304ec207d97c64fdb2bc11
Use volume o.vo in get_backup_volume_temp_snapshot

In snapshot's path in BaseVD's get_backup_device we're fetching a
volume using SQLAlchemy API instead of objects. In the end this caused
that sometimes we're returning a volume dictionary instead of an o.vo
from get_backup_device call in volume.manager. As in backup.manager
we're assuming that backup_device will be an object, backup operation
sometimes exploded.

This commit fixes that by making sure that we're always returning an
o.vo from that call.

Change-Id: I7adf3655d9c1bc63baede8f56202fd139709e552
Closes-Bug: 1607333