Updated openstack/openstack Project: openstack-infra/shade  f238253764544a7fbf2d418ef43c281c216e2771
Detect the need for FIPs better in auto_ip

The current logic of looking at interface_ip to see if a FIP is needed
is too simplistic and mixes two concerns. interface_ip is dependent on
the network context of the calling host, which is great - but may not be
the full story for all of the people who would want to interact with the
server.

Instead of creating a FIP if there is no interface_ip - introduce a
richer method that looks at public and private v4 addresses as well as
the existence of a FIP-able network.

Change-Id: I687375a1433d643e2f1cf18d0bbc00d0e8a0abbf