Updated openstack/openstack Project: openstack-infra/shade  176be18597076709225c62ab6c19312db2850158
Log request ids

We collect request ids from the cloud if they're there. Then, the
current approach is to attach them to the object - but they're ephemeral
and not actually a quality of the object - they're useful for debugging.
Instead of making an object property, log them. That way they can be
used where they're useful - ops logs for after the fact debugging.

The logger is named so that it can be explicitly controlled in a logging
config.

Change-Id: Ie5532cdc316cb00d5f15cae8b2a2f8674ae9da40